make input of diagnostics_flag_check simple and make large int value inputs robust
